Release SmartRouting
