+++ core/components/smartrouting/composer.json
@@ -1,0 +1,30 @@
+{
+  "name": "treehillstudio/smartrouting",
+  "description": "Switch the context based on a (sub)domain and/or subfolder",
+  "authors": [
+    {
+      "name": "Christian Seel",
+      "email": "office@seda.digital",
+      "homepage": "https://seda.digital/"
+    },
+    {
+      "name": "Thomas Jakobi",
+      "email": "office@treehillstudio.com",
+      "homepage": "https://treehillstudio.com"
+    }
+  ],
+  "license": "GPL-2.0-or-later",
+  "config": {
+    "platform": {
+      "php": "7.2"
+    }
+  },
+  "require": {
+    "php": ">=7.2"
+  },
+  "autoload": {
+    "psr-4": {
+      "TreehillStudio\\SmartRouting\\": "src/"
+    }
+  }
+}
