+++ core/components/smartrouting/docs/license.md
