+++ core/components/smartrouting/docs/readme.md
@@ -1,0 +1,26 @@
+# SmartRouting
+
+Switch the context based on a (sub)domain and/or subfolder.
+
+- Author: Thomas Jakobi <office@treehillstudio.com>
+- License: GNU GPLv2
+
+Based on the code of XRouting created by Christian Seel <office@seda.digital>
+
+# Features
+
+- Switch the context based on a (sub)domain AND/OR subfolder.
+- Define the routing based on the http_host and base_url context settings.
+- Cache the http_host/base_url => context settings to avoid database lookups on page load.
+
+## Installation
+
+MODX Package Management
+
+## Documentation
+
+For more information please read the documentation on https://jako.github.io/SmartRouting/
+
+## GitHub Repository
+
+https://github.com/Jako/SmartRouting
