+++ core/components/smartrouting/lexicon/en/default.inc.php
@@ -1,0 +1,8 @@
+<?php
+/**
+ * Default Lexicon Entries for SmartRouting
+ *
+ * @package smartrouting
+ * @subpackage lexicon
+ */
+$_lang['smartrouting'] = 'SmartRouting';
