+++ core/components/smartrouting/model/smartrouting/smartrouting.class.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * SmartRouting
+ *
+ * Copyright 2023 by Thomas Jakobi - <office@treehillstudio.com>
+ *
+ * Based on the code of XRouting created by Christian Seel <office@seda.digital>
+ *
+ * @package smartrouting
+ * @subpackage classfile
+ */
+
+require_once dirname(__DIR__, 2) . '/vendor/autoload.php';
+
+/**
+ * Class SmartRouting
+ */
+class SmartRouting extends \TreehillStudio\SmartRouting\SmartRouting
+{
+}
