+++ core/components/smartrouting/src/Plugins/Events/OnContextSave.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * @package smartrouting
+ * @subpackage plugin
+ */
+
+namespace TreehillStudio\SmartRouting\Plugins\Events;
+
+class OnContextSave extends OnSiteRefresh
+{
+}
