--- core/components/xrouting/docs/readme.md
@@ -1,24 +1,0 @@
-# XRouting
-
-Simple routing plugin that handles requests for different contexts.
-
-- Author: Christian Seel <office@seda.digital>
-- License: GNU GPLv2
-
-# Features
-
-- Switch the context based on a (sub)domain AND/OR subfolder.
-- Define the routing based on the http_host and base_url context settings.
-- Cache the http_host/base_url => context settings to avoid database lookups on page load.
-
-## Installation
-
-MODX Package Management
-
-## Documentation
-
-For more information please read the documentation on https://jako.github.io/XRouting/
-
-## GitHub Repository
-
-https://github.com/Jako/XRouting
