--- core/components/xrouting/lexicon/de/default.inc.php
@@ -1,8 +1,0 @@
-<?php
-/**
- * Default Lexicon Entries for XRouting
- *
- * @package xrouting
- * @subpackage lexicon
- */
-$_lang['xrouting'] = 'XRouting';
