--- core/components/xrouting/lexicon/de/setting.inc.php
@@ -1,15 +1,0 @@
-<?php
-/**
- * Settings Lexicon Entries for XRouting
- *
- * @package xrouting
- * @subpackage lexicon
- */
-$_lang['setting_xrouting.allow_debug_info'] = 'Debug-Info erlauben';
-$_lang['setting_xrouting.allow_debug_info_desc'] = 'Aktivieren Sie diese Einstellung um die Ausgabe von Debug-Informationen zu erlauben. ACHTUNG: Diese Einstellung sollte unbedingt nach dem Debugging wieder deaktiviert werden, da eine Vielzahl von Informationen preisgegeben werden über das Projekt!';
-$_lang['setting_xrouting.default_context'] = 'Standard-Kontext';
-$_lang['setting_xrouting.default_context_desc'] = 'Der Standard-Kontext zu dem weitergeleitet wird falls kein passender Kontext gefunden wurde und xrouting.show_no_match_error deaktiviert ist.';
-$_lang['setting_xrouting.include_www'] = 'WWW-Subdomain automatisch einfügen';
-$_lang['setting_xrouting.include_www_desc'] = 'Aktivieren Sie diese Einstallung, damit XRouting die www-Subdomain automatisch einfügt. Bitte definieren Sie Ihre http_hosts ohne www wenn diese Einstellung aktiviert ist.';
-$_lang['setting_xrouting.show_no_match_error'] = 'Fehlermeldung ausgeben';
-$_lang['setting_xrouting.show_no_match_error_desc'] = 'Wenn Sie diese Einstellung aktivieren, wird XRouting auf Ihre Fehlerseite weiterleiten (error_page setting) falls kein passender Kontext gefunden wurde.';
