--- core/components/xrouting/src/Plugins/Events/OnContextRemove.php
@@ -1,11 +1,0 @@
-<?php
-/**
- * @package xrouting
- * @subpackage plugin
- */
-
-namespace SEDAdigital\XRouting\Plugins\Events;
-
-class OnContextRemove extends OnSiteRefresh
-{
-}
