--- core/components/xrouting/src/Plugins/Events/OnContextSave.php
@@ -1,11 +1,0 @@
-<?php
-/**
- * @package xrouting
- * @subpackage plugin
- */
-
-namespace SEDAdigital\XRouting\Plugins\Events;
-
-class OnContextSave extends OnSiteRefresh
-{
-}
