+++ docs/contributors.md
@@ -1,6 +1,10 @@
-The XRouting project was started in 2014 by [Christian Seel](https://seda.digital)
-and is maintained and developed further since 2023 by [Thomas
-Jakobi](https://github.com/jako).
+The SmartRouting project was started in 2023 by [Thomas
+Jakobi](https://github.com/jako) as rewrite of the popular XRouting extra for
+MODX. 
 
-Many thanks to all who contributed, whether by creating pull requests,
-submitting bug reports, or donating.
+Unfortunately, this extra has not been updated in the MODX repository since 2015
+and many unfixed issues have arisen since then, including the lack of full PHP 8
+compatibility.
+
+Many thanks to all who contributed to XRouting and SmartRouting, whether by
+creating pull requests, submitting bug reports, or donating.
