+++ docs/installation.md
@@ -1,23 +1,23 @@
 ## Install from MODX Extras
 
-Search for XRouting in the Package Manager of your MODX installation and install it
+Search for SmartRouting in the Package Manager of your MODX installation and install it
 in there.
 
 ## Manual installation
 
 If you can't access the MODX Extras Repository in your MODX installation, you
-can manually install XRouting.
+can manually install SmartRouting.
 
-* Download the transport package from [MODX Extras](https://modx.com/extras/package/xrouting) (or one of the pre built transport packages in [_packages](https://github.com/Jako/xrouting/tree/master/_packages))
+* Download the transport package from [MODX Extras](https://modx.com/extras/package/smartrouting) (or one of the pre built transport packages in [_packages](https://github.com/Jako/smartrouting/tree/master/_packages))
 * Upload the zip file to your MODX installation's `core/packages` folder or upload it manually in the MODX Package Manager.
 * In the MODX Manager, navigate to the Package Manager page, and select 'Search locally for packages' from the dropdown button.
-* XRouting should now show up in the list of available packages. Click the corresponding 'Install' button and follow the instructions to complete the installation.
+* SmartRouting should now show up in the list of available packages. Click the corresponding 'Install' button and follow the instructions to complete the installation.
 
 ## Build it from source
 
 To build and install the package from source you could use [Git Package
 Management](https://github.com/TheBoxer/Git-Package-Management). The GitHub
-repository of xrouting contains a
-[config.json](https://github.com/Jako/xrouting/blob/master/_build/config.json)
-to build that package locally. Use this option, if you want to debug XRouting
+repository of smartrouting contains a
+[config.json](https://github.com/Jako/smartrouting/blob/master/_build/config.json)
+to build that package locally. Use this option, if you want to debug SmartRouting
 and/or contribute bugfixes and enhancements.
