+++ mkdocs.yml
@@ -1,12 +1,12 @@
-site_name: XRouting
+site_name: SmartRouting
 docs_dir: docs
 site_dir: _build/site
 edit_uri: tree/master/docs
 
-repo_name: 'Jako/xrouting'
-repo_url: https://github.com/Jako/xrouting
+repo_name: 'Jako/SmartRouting'
+repo_url: https://github.com/Jako/SmartRouting
 
-copyright: '&copy; 2014-2023 <a href="https://seda.digital">SEDA.digital</a>, 2023 <a href="https://github.com/Jako">Thomas Jakobi</a>'
+copyright: '&copy; 2023 <a href="https://github.com/Jako">Thomas Jakobi</a>'
 
 nav:
 - Description: 'index.md'
@@ -21,11 +21,11 @@ extra:
     site: 40
   social:
     - icon: fontawesome/brands/modx
-      link: https://modx.com/extras/package/xrouting
+      link: https://modx.com/extras/package/smartrouting
 
 theme:
   name: material
-  custom_dir: ../../mkdocs-material-jako/material
+  custom_dir: ../../mkdocs-material-treehillstudio/material
 
   # 404 page
   static_templates:
